pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-market-maker"
version = "0.1.0"
description = "Kalshi market making bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests==2.26.0",
  "python-dotenv==0.19.1",
  "PyYAML==6.0",
  "cryptography==44.0.1",
]

[project.scripts]
kalshi-mm = "kalshi_market_maker.cli.mm:main"
kalshi-cancel-all = "kalshi_market_maker.cli.cancel_all:main"
kalshi-dashboard = "kalshi_market_maker.cli.dashboard:main"

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["kalshi_market_maker*"]
